Minor bugfixes after cherry-picking the o11yphant tracing refactors over
