+++ filers/default/src/main/java/org/commonjava/indy/filer/def/DefaultGalleyStorageProvider.java
@@ -402,6 +402,7 @@ public class DefaultGalleyStorageProvider
             catch ( GalleyInitException e )
             {
                 logger.error( "[Indy] Can not create CacheProvider for some error.", e );
+                throw new RuntimeException( "Cannot create CacheProvider", e );
             }
         }
