+++ pom.xml
@@ -65,6 +65,8 @@
     <protostreamVersion>4.3.0.Final</protostreamVersion>
     <gsonVersion>2.8.2</gsonVersion>
     <!--<luceneVersion>7.3.0</luceneVersion>-->
+    <protostreamVersion>4.3.0.Final</protostreamVersion>
+    <gsonVersion>2.8.2</gsonVersion>
     <hibernateVersion>5.8.1.Final</hibernateVersion>
     <resteasyVersion>3.0.12.Final</resteasyVersion>
     <undertowVersion>2.1.0.Final</undertowVersion>
@@ -402,12 +404,12 @@
       <dependency>
         <groupId>org.commonjava.indy</groupId>
         <artifactId>indy-subsys-trace</artifactId>
-        <version>2.6.0-SNAPSHOT</version>
+        <version>2.5.5-SNAPSHOT</version>
       </dependency>
       <dependency>
         <groupId>org.commonjava.indy</groupId>
         <artifactId>indy-subsys-trace</artifactId>
-        <version>2.6.0-SNAPSHOT</version>
+        <version>2.5.5-SNAPSHOT</version>
         <type>tar.gz</type>
         <classifier>confset</classifier>
         <scope>provided</scope>
