+++ subsys/trace/pom.xml
@@ -60,10 +60,6 @@
       <groupId>org.commonjava.indy</groupId>
       <artifactId>indy-subsys-metrics</artifactId>
     </dependency>
-    <dependency>
-      <groupId>org.commonjava.indy</groupId>
-      <artifactId>indy-subsys-cassandra</artifactId>
-    </dependency>
   </dependencies>
 
   <build>
